Update the tag name for uberfire due to an issue with the previous tag
A problem was found with the https://github.com/jboss-integration/uberfire/tree/sync-6.3.x-2015.11.26 Uberfire tag after it was published. A new tag was created fixing the problem and re-published. The new tag is available at https://github.com/jboss-integration/uberfire/tree/sync-6.3.x-2015.11.26-2 . The report was amended to include this information.